pyproject.toml
[project]
name = "ag2"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=45.0.4",
    "flet[all]>=0.28.3",
    "mysql-connector-python>=9.3.0",
    "notebook>=7.4.2",
    "pandas>=2.2.3",
    "pandas-stubs==2.2.3.250527",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "sqlalchemy>=2.0.41",
]
